Revert "Get prop types from inner component"
