--- mocks/ComponentAST2.js
@@ -1,1164 +1,0 @@
-[
-  {
-    "type": "File",
-    "start": 0,
-    "end": 467,
-    "loc": {
-      "start": {
-        "line": 1,
-        "column": 0
-      },
-      "end": {
-        "line": 24,
-        "column": 0
-      }
-    },
-    "program": {
-      "type": "Program",
-      "start": 0,
-      "end": 467,
-      "loc": {
-        "start": {
-          "line": 1,
-          "column": 0
-        },
-        "end": {
-          "line": 24,
-          "column": 0
-        }
-      },
-      "sourceType": "module",
-      "interpreter": null,
-      "body": [
-        {
-          "type": "ImportDeclaration",
-          "start": 0,
-          "end": 26,
-          "loc": {
-            "start": {
-              "line": 1,
-              "column": 0
-            },
-            "end": {
-              "line": 1,
-              "column": 26
-            }
-          },
-          "specifiers": [
-            {
-              "type": "ImportDefaultSpecifier",
-              "start": 7,
-              "end": 12,
-              "loc": {
-                "start": {
-                  "line": 1,
-                  "column": 7
-                },
-                "end": {
-                  "line": 1,
-                  "column": 12
-                }
-              },
-              "local": {
-                "type": "Identifier",
-                "start": 7,
-                "end": 12,
-                "loc": {
-                  "start": {
-                    "line": 1,
-                    "column": 7
-                  },
-                  "end": {
-                    "line": 1,
-                    "column": 12
-                  },
-                  "identifierName": "React"
-                },
-                "name": "React"
-              }
-            }
-          ],
-          "source": {
-            "type": "StringLiteral",
-            "start": 18,
-            "end": 25,
-            "loc": {
-              "start": {
-                "line": 1,
-                "column": 18
-              },
-              "end": {
-                "line": 1,
-                "column": 25
-              }
-            },
-            "extra": {
-              "rawValue": "react",
-              "raw": "'react'"
-            },
-            "value": "react"
-          }
-        },
-        {
-          "type": "ImportDeclaration",
-          "start": 27,
-          "end": 62,
-          "loc": {
-            "start": {
-              "line": 2,
-              "column": 0
-            },
-            "end": {
-              "line": 2,
-              "column": 35
-            }
-          },
-          "specifiers": [
-            {
-              "type": "ImportDefaultSpecifier",
-              "start": 34,
-              "end": 43,
-              "loc": {
-                "start": {
-                  "line": 2,
-                  "column": 7
-                },
-                "end": {
-                  "line": 2,
-                  "column": 16
-                }
-              },
-              "local": {
-                "type": "Identifier",
-                "start": 34,
-                "end": 43,
-                "loc": {
-                  "start": {
-                    "line": 2,
-                    "column": 7
-                  },
-                  "end": {
-                    "line": 2,
-                    "column": 16
-                  },
-                  "identifierName": "PropTypes"
-                },
-                "name": "PropTypes"
-              }
-            }
-          ],
-          "source": {
-            "type": "StringLiteral",
-            "start": 49,
-            "end": 61,
-            "loc": {
-              "start": {
-                "line": 2,
-                "column": 22
-              },
-              "end": {
-                "line": 2,
-                "column": 34
-              }
-            },
-            "extra": {
-              "rawValue": "prop-types",
-              "raw": "'prop-types'"
-            },
-            "value": "prop-types"
-          }
-        },
-        {
-          "type": "ImportDeclaration",
-          "start": 63,
-          "end": 111,
-          "loc": {
-            "start": {
-              "line": 3,
-              "column": 0
-            },
-            "end": {
-              "line": 3,
-              "column": 48
-            }
-          },
-          "specifiers": [
-            {
-              "type": "ImportSpecifier",
-              "start": 72,
-              "end": 85,
-              "loc": {
-                "start": {
-                  "line": 3,
-                  "column": 9
-                },
-                "end": {
-                  "line": 3,
-                  "column": 22
-                }
-              },
-              "imported": {
-                "type": "Identifier",
-                "start": 72,
-                "end": 85,
-                "loc": {
-                  "start": {
-                    "line": 3,
-                    "column": 9
-                  },
-                  "end": {
-                    "line": 3,
-                    "column": 22
-                  },
-                  "identifierName": "TestComponent"
-                },
-                "name": "TestComponent"
-              },
-              "local": {
-                "type": "Identifier",
-                "start": 72,
-                "end": 85,
-                "loc": {
-                  "start": {
-                    "line": 3,
-                    "column": 9
-                  },
-                  "end": {
-                    "line": 3,
-                    "column": 22
-                  },
-                  "identifierName": "TestComponent"
-                },
-                "name": "TestComponent"
-              }
-            }
-          ],
-          "source": {
-            "type": "StringLiteral",
-            "start": 93,
-            "end": 110,
-            "loc": {
-              "start": {
-                "line": 3,
-                "column": 30
-              },
-              "end": {
-                "line": 3,
-                "column": 47
-              }
-            },
-            "extra": {
-              "rawValue": "./TestComponent",
-              "raw": "'./TestComponent'"
-            },
-            "value": "./TestComponent"
-          }
-        },
-        {
-          "type": "ImportDeclaration",
-          "start": 112,
-          "end": 162,
-          "loc": {
-            "start": {
-              "line": 4,
-              "column": 0
-            },
-            "end": {
-              "line": 4,
-              "column": 50
-            }
-          },
-          "specifiers": [
-            {
-              "type": "ImportSpecifier",
-              "start": 121,
-              "end": 135,
-              "loc": {
-                "start": {
-                  "line": 4,
-                  "column": 9
-                },
-                "end": {
-                  "line": 4,
-                  "column": 23
-                }
-              },
-              "imported": {
-                "type": "Identifier",
-                "start": 121,
-                "end": 135,
-                "loc": {
-                  "start": {
-                    "line": 4,
-                    "column": 9
-                  },
-                  "end": {
-                    "line": 4,
-                    "column": 23
-                  },
-                  "identifierName": "OtherComponent"
-                },
-                "name": "OtherComponent"
-              },
-              "local": {
-                "type": "Identifier",
-                "start": 121,
-                "end": 135,
-                "loc": {
-                  "start": {
-                    "line": 4,
-                    "column": 9
-                  },
-                  "end": {
-                    "line": 4,
-                    "column": 23
-                  },
-                  "identifierName": "OtherComponent"
-                },
-                "name": "OtherComponent"
-              }
-            }
-          ],
-          "source": {
-            "type": "StringLiteral",
-            "start": 143,
-            "end": 161,
-            "loc": {
-              "start": {
-                "line": 4,
-                "column": 31
-              },
-              "end": {
-                "line": 4,
-                "column": 49
-              }
-            },
-            "extra": {
-              "rawValue": "./OtherComponent",
-              "raw": "'./OtherComponent'"
-            },
-            "value": "./OtherComponent"
-          }
-        },
-        {
-          "type": "VariableDeclaration",
-          "start": 164,
-          "end": 385,
-          "loc": {
-            "start": {
-              "line": 6,
-              "column": 0
-            },
-            "end": {
-              "line": 17,
-              "column": 2
-            }
-          },
-          "declarations": [
-            {
-              "type": "VariableDeclarator",
-              "start": 170,
-              "end": 384,
-              "loc": {
-                "start": {
-                  "line": 6,
-                  "column": 6
-                },
-                "end": {
-                  "line": 17,
-                  "column": 1
-                }
-              },
-              "id": {
-                "type": "Identifier",
-                "start": 170,
-                "end": 179,
-                "loc": {
-                  "start": {
-                    "line": 6,
-                    "column": 6
-                  },
-                  "end": {
-                    "line": 6,
-                    "column": 15
-                  },
-                  "identifierName": "Component"
-                },
-                "name": "Component"
-              },
-              "init": {
-                "type": "ArrowFunctionExpression",
-                "start": 182,
-                "end": 384,
-                "loc": {
-                  "start": {
-                    "line": 6,
-                    "column": 18
-                  },
-                  "end": {
-                    "line": 17,
-                    "column": 1
-                  }
-                },
-                "id": null,
-                "generator": false,
-                "async": false,
-                "params": [],
-                "body": {
-                  "type": "JSXElement",
-                  "start": 192,
-                  "end": 382,
-                  "loc": {
-                    "start": {
-                      "line": 7,
-                      "column": 2
-                    },
-                    "end": {
-                      "line": 16,
-                      "column": 18
-                    }
-                  },
-                  "openingElement": {
-                    "type": "JSXOpeningElement",
-                    "start": 192,
-                    "end": 340,
-                    "loc": {
-                      "start": {
-                        "line": 7,
-                        "column": 2
-                      },
-                      "end": {
-                        "line": 14,
-                        "column": 3
-                      }
-                    },
-                    "name": {
-                      "type": "JSXIdentifier",
-                      "start": 193,
-                      "end": 206,
-                      "loc": {
-                        "start": {
-                          "line": 7,
-                          "column": 3
-                        },
-                        "end": {
-                          "line": 7,
-                          "column": 16
-                        }
-                      },
-                      "name": "TestComponent"
-                    },
-                    "attributes": [
-                      {
-                        "type": "JSXAttribute",
-                        "start": 211,
-                        "end": 224,
-                        "loc": {
-                          "start": {
-                            "line": 8,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 8,
-                            "column": 17
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 211,
-                          "end": 215,
-                          "loc": {
-                            "start": {
-                              "line": 8,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 8,
-                              "column": 8
-                            }
-                          },
-                          "name": "name"
-                        },
-                        "value": {
-                          "type": "StringLiteral",
-                          "start": 216,
-                          "end": 224,
-                          "loc": {
-                            "start": {
-                              "line": 8,
-                              "column": 9
-                            },
-                            "end": {
-                              "line": 8,
-                              "column": 17
-                            }
-                          },
-                          "extra": {
-                            "rawValue": "string",
-                            "raw": "'string'"
-                          },
-                          "value": "string"
-                        }
-                      },
-                      {
-                        "type": "JSXAttribute",
-                        "start": 229,
-                        "end": 242,
-                        "loc": {
-                          "start": {
-                            "line": 9,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 9,
-                            "column": 17
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 229,
-                          "end": 235,
-                          "loc": {
-                            "start": {
-                              "line": 9,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 9,
-                              "column": 10
-                            }
-                          },
-                          "name": "isReal"
-                        },
-                        "value": {
-                          "type": "JSXExpressionContainer",
-                          "start": 236,
-                          "end": 242,
-                          "loc": {
-                            "start": {
-                              "line": 9,
-                              "column": 11
-                            },
-                            "end": {
-                              "line": 9,
-                              "column": 17
-                            }
-                          },
-                          "expression": {
-                            "type": "BooleanLiteral",
-                            "start": 237,
-                            "end": 241,
-                            "loc": {
-                              "start": {
-                                "line": 9,
-                                "column": 12
-                              },
-                              "end": {
-                                "line": 9,
-                                "column": 16
-                              }
-                            },
-                            "value": true
-                          }
-                        }
-                      },
-                      {
-                        "type": "JSXAttribute",
-                        "start": 247,
-                        "end": 259,
-                        "loc": {
-                          "start": {
-                            "line": 10,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 10,
-                            "column": 16
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 247,
-                          "end": 255,
-                          "loc": {
-                            "start": {
-                              "line": 10,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 10,
-                              "column": 12
-                            }
-                          },
-                          "name": "bedrooms"
-                        },
-                        "value": {
-                          "type": "JSXExpressionContainer",
-                          "start": 256,
-                          "end": 259,
-                          "loc": {
-                            "start": {
-                              "line": 10,
-                              "column": 13
-                            },
-                            "end": {
-                              "line": 10,
-                              "column": 16
-                            }
-                          },
-                          "expression": {
-                            "type": "NumericLiteral",
-                            "start": 257,
-                            "end": 258,
-                            "loc": {
-                              "start": {
-                                "line": 10,
-                                "column": 14
-                              },
-                              "end": {
-                                "line": 10,
-                                "column": 15
-                              }
-                            },
-                            "extra": {
-                              "rawValue": 2,
-                              "raw": "2"
-                            },
-                            "value": 2
-                          }
-                        }
-                      },
-                      {
-                        "type": "JSXAttribute",
-                        "start": 264,
-                        "end": 274,
-                        "loc": {
-                          "start": {
-                            "line": 11,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 11,
-                            "column": 14
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 264,
-                          "end": 269,
-                          "loc": {
-                            "start": {
-                              "line": 11,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 11,
-                              "column": 9
-                            }
-                          },
-                          "name": "width"
-                        },
-                        "value": {
-                          "type": "StringLiteral",
-                          "start": 270,
-                          "end": 274,
-                          "loc": {
-                            "start": {
-                              "line": 11,
-                              "column": 10
-                            },
-                            "end": {
-                              "line": 11,
-                              "column": 14
-                            }
-                          },
-                          "extra": {
-                            "rawValue": "xs",
-                            "raw": "'xs'"
-                          },
-                          "value": "xs"
-                        }
-                      },
-                      {
-                        "type": "JSXAttribute",
-                        "start": 279,
-                        "end": 310,
-                        "loc": {
-                          "start": {
-                            "line": 12,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 12,
-                            "column": 35
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 279,
-                          "end": 293,
-                          "loc": {
-                            "start": {
-                              "line": 12,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 12,
-                              "column": 18
-                            }
-                          },
-                          "name": "breadcrumbInfo"
-                        },
-                        "value": {
-                          "type": "JSXExpressionContainer",
-                          "start": 294,
-                          "end": 310,
-                          "loc": {
-                            "start": {
-                              "line": 12,
-                              "column": 19
-                            },
-                            "end": {
-                              "line": 12,
-                              "column": 35
-                            }
-                          },
-                          "expression": {
-                            "type": "Identifier",
-                            "start": 295,
-                            "end": 309,
-                            "loc": {
-                              "start": {
-                                "line": 12,
-                                "column": 20
-                              },
-                              "end": {
-                                "line": 12,
-                                "column": 34
-                              },
-                              "identifierName": "breadcrumbInfo"
-                            },
-                            "name": "breadcrumbInfo"
-                          }
-                        }
-                      },
-                      {
-                        "type": "JSXAttribute",
-                        "start": 315,
-                        "end": 336,
-                        "loc": {
-                          "start": {
-                            "line": 13,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 13,
-                            "column": 25
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 315,
-                          "end": 324,
-                          "loc": {
-                            "start": {
-                              "line": 13,
-                              "column": 4
-                            },
-                            "end": {
-                              "line": 13,
-                              "column": 13
-                            }
-                          },
-                          "name": "houseInfo"
-                        },
-                        "value": {
-                          "type": "JSXExpressionContainer",
-                          "start": 325,
-                          "end": 336,
-                          "loc": {
-                            "start": {
-                              "line": 13,
-                              "column": 14
-                            },
-                            "end": {
-                              "line": 13,
-                              "column": 25
-                            }
-                          },
-                          "expression": {
-                            "type": "Identifier",
-                            "start": 326,
-                            "end": 335,
-                            "loc": {
-                              "start": {
-                                "line": 13,
-                                "column": 15
-                              },
-                              "end": {
-                                "line": 13,
-                                "column": 24
-                              },
-                              "identifierName": "houseInfo"
-                            },
-                            "name": "houseInfo"
-                          }
-                        }
-                      }
-                    ],
-                    "selfClosing": false
-                  },
-                  "closingElement": {
-                    "type": "JSXClosingElement",
-                    "start": 366,
-                    "end": 382,
-                    "loc": {
-                      "start": {
-                        "line": 16,
-                        "column": 2
-                      },
-                      "end": {
-                        "line": 16,
-                        "column": 18
-                      }
-                    },
-                    "name": {
-                      "type": "JSXIdentifier",
-                      "start": 368,
-                      "end": 381,
-                      "loc": {
-                        "start": {
-                          "line": 16,
-                          "column": 4
-                        },
-                        "end": {
-                          "line": 16,
-                          "column": 17
-                        }
-                      },
-                      "name": "TestComponent"
-                    }
-                  },
-                  "children": [
-                    {
-                      "type": "JSXText",
-                      "start": 340,
-                      "end": 345,
-                      "loc": {
-                        "start": {
-                          "line": 14,
-                          "column": 3
-                        },
-                        "end": {
-                          "line": 15,
-                          "column": 4
-                        }
-                      },
-                      "extra": {
-                        "rawValue": "\n    ",
-                        "raw": "\n    "
-                      },
-                      "value": "\n    "
-                    },
-                    {
-                      "type": "JSXElement",
-                      "start": 345,
-                      "end": 363,
-                      "loc": {
-                        "start": {
-                          "line": 15,
-                          "column": 4
-                        },
-                        "end": {
-                          "line": 15,
-                          "column": 22
-                        }
-                      },
-                      "openingElement": {
-                        "type": "JSXOpeningElement",
-                        "start": 345,
-                        "end": 363,
-                        "loc": {
-                          "start": {
-                            "line": 15,
-                            "column": 4
-                          },
-                          "end": {
-                            "line": 15,
-                            "column": 22
-                          }
-                        },
-                        "name": {
-                          "type": "JSXIdentifier",
-                          "start": 346,
-                          "end": 360,
-                          "loc": {
-                            "start": {
-                              "line": 15,
-                              "column": 5
-                            },
-                            "end": {
-                              "line": 15,
-                              "column": 19
-                            }
-                          },
-                          "name": "OtherComponent"
-                        },
-                        "attributes": [],
-                        "selfClosing": true
-                      },
-                      "closingElement": null,
-                      "children": []
-                    },
-                    {
-                      "type": "JSXText",
-                      "start": 363,
-                      "end": 366,
-                      "loc": {
-                        "start": {
-                          "line": 15,
-                          "column": 22
-                        },
-                        "end": {
-                          "line": 16,
-                          "column": 2
-                        }
-                      },
-                      "extra": {
-                        "rawValue": "\n  ",
-                        "raw": "\n  "
-                      },
-                      "value": "\n  "
-                    }
-                  ],
-                  "extra": {
-                    "parenthesized": true,
-                    "parenStart": 188
-                  }
-                }
-              }
-            }
-          ],
-          "kind": "const"
-        },
-        {
-          "type": "ExpressionStatement",
-          "start": 387,
-          "end": 439,
-          "loc": {
-            "start": {
-              "line": 19,
-              "column": 0
-            },
-            "end": {
-              "line": 21,
-              "column": 2
-            }
-          },
-          "expression": {
-            "type": "AssignmentExpression",
-            "start": 387,
-            "end": 438,
-            "loc": {
-              "start": {
-                "line": 19,
-                "column": 0
-              },
-              "end": {
-                "line": 21,
-                "column": 1
-              }
-            },
-            "operator": "=",
-            "left": {
-              "type": "MemberExpression",
-              "start": 387,
-              "end": 406,
-              "loc": {
-                "start": {
-                  "line": 19,
-                  "column": 0
-                },
-                "end": {
-                  "line": 19,
-                  "column": 19
-                }
-              },
-              "object": {
-                "type": "Identifier",
-                "start": 387,
-                "end": 396,
-                "loc": {
-                  "start": {
-                    "line": 19,
-                    "column": 0
-                  },
-                  "end": {
-                    "line": 19,
-                    "column": 9
-                  },
-                  "identifierName": "Component"
-                },
-                "name": "Component"
-              },
-              "property": {
-                "type": "Identifier",
-                "start": 397,
-                "end": 406,
-                "loc": {
-                  "start": {
-                    "line": 19,
-                    "column": 10
-                  },
-                  "end": {
-                    "line": 19,
-                    "column": 19
-                  },
-                  "identifierName": "propTypes"
-                },
-                "name": "propTypes"
-              },
-              "computed": false
-            },
-            "right": {
-              "type": "ObjectExpression",
-              "start": 409,
-              "end": 438,
-              "loc": {
-                "start": {
-                  "line": 19,
-                  "column": 22
-                },
-                "end": {
-                  "line": 21,
-                  "column": 1
-                }
-              },
-              "properties": [
-                {
-                  "type": "ObjectProperty",
-                  "start": 413,
-                  "end": 435,
-                  "loc": {
-                    "start": {
-                      "line": 20,
-                      "column": 2
-                    },
-                    "end": {
-                      "line": 20,
-                      "column": 24
-                    }
-                  },
-                  "method": false,
-                  "key": {
-                    "type": "Identifier",
-                    "start": 413,
-                    "end": 417,
-                    "loc": {
-                      "start": {
-                        "line": 20,
-                        "column": 2
-                      },
-                      "end": {
-                        "line": 20,
-                        "column": 6
-                      },
-                      "identifierName": "name"
-                    },
-                    "name": "name"
-                  },
-                  "computed": false,
-                  "shorthand": false,
-                  "value": {
-                    "type": "MemberExpression",
-                    "start": 419,
-                    "end": 435,
-                    "loc": {
-                      "start": {
-                        "line": 20,
-                        "column": 8
-                      },
-                      "end": {
-                        "line": 20,
-                        "column": 24
-                      }
-                    },
-                    "object": {
-                      "type": "Identifier",
-                      "start": 419,
-                      "end": 428,
-                      "loc": {
-                        "start": {
-                          "line": 20,
-                          "column": 8
-                        },
-                        "end": {
-                          "line": 20,
-                          "column": 17
-                        },
-                        "identifierName": "PropTypes"
-                      },
-                      "name": "PropTypes"
-                    },
-                    "property": {
-                      "type": "Identifier",
-                      "start": 429,
-                      "end": 435,
-                      "loc": {
-                        "start": {
-                          "line": 20,
-                          "column": 18
-                        },
-                        "end": {
-                          "line": 20,
-                          "column": 24
-                        },
-                        "identifierName": "string"
-                      },
-                      "name": "string"
-                    },
-                    "computed": false
-                  }
-                }
-              ]
-            }
-          }
-        },
-        {
-          "type": "ExportDefaultDeclaration",
-          "start": 441,
-          "end": 466,
-          "loc": {
-            "start": {
-              "line": 23,
-              "column": 0
-            },
-            "end": {
-              "line": 23,
-              "column": 25
-            }
-          },
-          "declaration": {
-            "type": "Identifier",
-            "start": 456,
-            "end": 465,
-            "loc": {
-              "start": {
-                "line": 23,
-                "column": 15
-              },
-              "end": {
-                "line": 23,
-                "column": 24
-              },
-              "identifierName": "Component"
-            },
-            "name": "Component"
-          }
-        }
-      ],
-      "directives": []
-    },
-    "comments": []
-  }
-]
