--- mocks/OtherComponent.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export class OtherComponent extends React.PureComponent {
-  render() {
-    return <div />;
-  }
-}
-
-export default OtherComponent;
