+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "boring-test",
-  "version": "1.7.0-alpha.0",
+  "version": "1.6.3",
   "description": "Automate boring tests",
   "main": "index.js",
   "repository": "git@github.com:leandrotk/boring-test.git",
@@ -8,16 +8,14 @@
   "license": "MIT",
   "scripts": {
     "test": "jest",
-    "test:watch": "jest --no-cache --watch",
+    "watch": "jest --no-cache --watch ./*",
     "build": "babel src -d lib",
     "build:watch": "babel --watch src -d lib",
-    "build:run": "babel src -d lib && node lib/index.js mocks/Component.js -t ui",
+    "build:run": "babel src -d lib && node lib/index.js src/core.js -t css",
     "build:generate": "babel src -d lib && node lib/index.js generate-templates"
   },
   "dependencies": {
-    "@babel/parser": "^7.6.2",
     "@babel/runtime": "^7.6.0",
-    "@babel/traverse": "^7.6.3",
     "minimist": "^1.2.0",
     "mkdirp": "^0.5.1"
   },
