--- src/propTypes/tests/getComponentsFilePaths.test.js
@@ -1,14 +1,0 @@
-import { getComponentsFilePaths } from '../getComponentsFilePaths';
-
-describe('getComponentsFilePaths', () => {
-  it('generates proper path from component file', async () => {
-    const filePath = 'mocks/Component.js';
-
-    const paths = await getComponentsFilePaths(filePath);
-
-    expect(paths).toMatchObject({
-      OtherComponent: "./OtherComponent",
-      TestComponent: "./TestComponent"
-    })
-  });
-});
