+++ test/read.test.js
@@ -1,7 +1,8 @@
 import {
   getTemplateFile,
   getComponentName,
-} from "../read";
+  componentReplacement
+} from "../src/read";
 
 describe("getTemplateFile", () => {
   describe("with no template file", () => {
@@ -22,5 +23,14 @@ describe("getComponentName", () => {
   it("returns the component name from the file", () => {
     const filePath = "app/containers/Component.js";
     expect(getComponentName(filePath)).toEqual("Component");
+  });
+});
+
+describe("componentReplacement", () => {
+  it("returns the component name from the file", () => {
+    const template = "<${componentToBeTested}>Some</${componentToBeTested}>";
+    const component = "Component";
+    const newTest = `<${component}>Some</${component}>`;
+    expect(componentReplacement(template, component)).toEqual(newTest);
   });
 });
