--- utils/logger.js
@@ -1,2 +1,0 @@
-const logObject = ({ message, value }) =>
-  console.log(message, JSON.stringify(value, null, 2));
